+++ src/components/form/Anexos.vue
@@ -70,22 +70,27 @@
               </button>
             </div>
           </div>
-          <div class="flex items-center gap-2" v-if="!isViewOnly">
+          <div class="flex items-center gap-2">
+            <div class="flex items-center gap-2" v-if="!isViewOnly">
+              <button
+                type="button"
+                v-if="!anexo.isEditing"
+                @click="startEditing(anexo)"
+                class="text-gray-500 hover:text-gray-700 transition duration-200 ease-in-out transform hover:scale-110 text-base text-[1.25rem] mt-[2px]"
+              >
+                ✏️
+              </button>
+              <button
+                type="button"
+                v-if="!anexo.isEditing"
+                @click="deleteFile(anexo.id)"
+                class="text-[#e74c3c] hover:text-[#c0392b] transition duration-200 ease-in-out transform hover:scale-110 text-[2.5rem] mt-[2px]"
+              >
+                &times;
+              </button>
+          </div>
             <button
-              v-if="!anexo.isEditing"
-              @click="startEditing(anexo)"
-              class="text-gray-500 hover:text-gray-700 transition duration-200 ease-in-out transform hover:scale-110 text-base text-[1.25rem] mt-[2px]"
-            >
-              ✏️
-            </button>
-            <button
-              v-if="!anexo.isEditing"
-              @click="deleteFile(anexo.id)"
-              class="text-[#e74c3c] hover:text-[#c0392b] transition duration-200 ease-in-out transform hover:scale-110 text-[2.5rem] mt-[2px]"
-            >
-              &times;
-            </button>
-            <button
+              type="button"
               v-if="!anexo.isEditing"
               @click.prevent="downloadAnexo(anexo)"
               title="Download"
@@ -105,6 +110,7 @@
 import { ref, onMounted } from 'vue';
 import { api } from '@/services/api';
 import { Icon } from '@iconify/vue';
+import Swal from 'sweetalert2';
 
 const selectedFiles = ref(null);
 const anexos = ref([]);
@@ -210,18 +216,33 @@ const fetchAnexos = async () => {
 };
 
 const deleteFile = async (id) => {
-  try {
-    let variantUrl = props.variant === 'contrato' ? 'contratos' : props.variant === 'medicao' ? 'medicao' : 'faturamento';
+  const result = await Swal.fire({
+    title: 'Tem certeza?',
+    text: 'Esta ação irá deletar o anexo permanentemente.',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#e74c3c',
+    cancelButtonColor: '#3498db',
+    confirmButtonText: 'Sim, deletar!',
+    cancelButtonText: 'Cancelar',
+  }).then(async (result)=>{
+    if (result.isConfirmed){
+      try {
+        let variantUrl = props.variant === 'contrato' ? 'contratos' : props.variant === 'medicao' ? 'medicao' : 'faturamento';
 
-    await api.delete(`/${variantUrl}/anexos/${id}`);
-    successMessage.value = 'Arquivo excluído com sucesso!';
-    fetchAnexos();
-    setTimeout(() => {
-      successMessage.value = '';
-    }, 3000);
-  } catch (error) {
-    errorMessage.value = 'Erro ao excluir o arquivo.';
-  }
+        await api.delete(`/${variantUrl}/anexos/${id}`);
+        successMessage.value = 'Arquivo excluído com sucesso!';
+        fetchAnexos();
+        setTimeout(() => {
+          successMessage.value = '';
+        }, 3000);
+      } catch (error) {
+        errorMessage.value = 'Erro ao excluir o arquivo.';
+      }
+    }
+  })
+
+
 };
 
 const downloadAnexo = async (anexo) => {
